pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "harness"
version = "3.0.0"
description = "Spec-governed, risk-classified engineering CLI for AI agents"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
harness = "harness_runtime.cli:main"

[tool.setuptools.packages.find]
include = ["harness_runtime*"]

[tool.setuptools.package-data]
harness_runtime = ["../resources/policies/*.yaml"]
